Added support for security contexts and pluggable protocols
Workflow inputs can come from different sources, and they can be complex
ones like EGA, for instance. So, security contexts related to inputs
help on that task, separating the credentials from the URL.

Also, input materialization has been revamped, so it supports pluggable
protocols and security contexts. In order to avoid collisions at the
filename level, a hash is used instead of the derived filename. Detailed
and linearized information about parameters and materialized files is
gathered, in order to be used when the workflow is going to be run.

Last, in order to avoid a possible collision of a "class" parameter with
the input class keyword, the keyword is now 'c-l-a-s-s'.