Fixed regression introduced on commit e2032a0faf2b35a85d83aa57794e00234c496a6d
When stage unmarshalling was enabled, a regression related to the
regeneration of MaterializedWorkflowEngine instance happened, as the
instance needs to contains the path to the patched / packed workflow.

In the case of CWL, the packed workflow is parsed to find implicit
inputs and declared outputs. If the workflow is not packed, it doesn't
work, and it fails in an unexpected place.