Unmarshalling now search for fixes in classes to be unmarshalled.
When an instance of a namedtuple or a class is unmarshalled, depending
on the divergence of the data structures it could happen that it fails
due new fields, renamed fields or removed fields.

The unmarshalling code now searches for class methods which can provide
hints about how to fix these issues.

Also, the default globals to use when classes have to be instantiated
now is the caller's one, not the one from the unmarshalling code.