Detect platform OS and processor architecture.
Both platform OS and processor architecture, provided by
basic-installer.bash, are used to download the most suitable
compiled version of OpenJDK, static bash and gocryptfs.

In this last case, if there is no available pre-compiled version,
Go is set up, gocryptfs source code is downloaded and then
compiled using either the available or the downloaded Go compiler.

It is still pending to get a pre-compiled version of either busybox
or ps command for any of the tested processor architectures.