Fixed issue related to new groovy parser package, realized by both @dcl10 and @paulaidt .
Version 0.1.0 of groovy-parser package was not properly built, so it did not
include the list of dependencies, and so the built wheel was incomplete.

New integrated version fixes those issues, and also introduces a caching
feature, needed because groovy parsing is quite slow right now.

The cached parsed contents are stored in the cache directory, under
`wf-cache/groovy-parsing-cache`.