Added hooks to parse, serialize and deserialize environment variables needed for workflow execution.
Also, added needed hooks to integrate the environment variables when a workflow is being run. Specific implementations for Nextflow and cwltool are added later in separate commits.

Last, but not the least important, environment variables are described in the generated RO-Crates as formal parameters. There is no standardized way to distinguish an environment variable from a classical parameter in the current representation.