Now a consolidated workflow directory is created inside the staged working directory.
In the past, consolidated workflows were kept only in the cache, so it was making very difficult to move those staged working directories to other locations.

Also, consolidated workflows were represented using an absolute path instead of a relative one in CWL, so it was a source of issues, like the one unconvered in the reworked RO-Crate generation code from several weeks ago.