Removed code duplication between podman_container and docker_container.
This work is needed to improve maintainability of both, as they
share the same code to compress saved images on the fly,
and decompress on the fly container images being loaded into the
local registry.

Also, switched container image compression from lzma to pgzip.
Images are bigger, but the parallelized version of gzip library
is much faster.