Rearranged the software requirements of a workflow.
As workflows can be run in very disparate environments with different
requisites, they only truly need the workflow engine as such to be run.

On the other hand, a CreateAction describing a workflow execution needs
to declare as instruments all the needed dependencies, which are WfExS-backend
itself, the workflow, the workflow engine, the container engine and
the list of needed software containers.