Fixes in serialization of FormalParameter and values in RO-Crate
Now, in the generated RO-Crate, the parameters which were explicitly set have
`requiredValue` as `True` in its FormalParameter declaration, otherwise it is `False`.

All the atomic values are now stringified.

Also, when the default value of a parameter could be recovered (right now, only with unset parameters),
the value is also set as `defaultValue` of the FormalParameter declaration.