Added proper RO-Crate description for ContentWithURIs.
ContentWithURIs are common use cases in life sciences workflows,
where an input file contains URIs to other contents to be fetched and
processed. In these cases, a file where the URIs are substituted by
local paths pointing to the very same content as the URIs would also work.

In order to assure WfExS input requirements are fetched before execution,
these kind of inputs have to be properly processed.