Added `README.md` to every generated RO-Crate.
As @simleo indicated, the not so common URI schemes which can appear in WfExS-backend RO-Crates
can be confusing, so an automatically generated `README.md` file describing the
possible CURIEs and URIs schemes is added to every generated RO-Crate.

A few schemes supported by `singularity` are still outside this automatic mechanism.