Zenodo export plugin now honors provided licences
Zenodo is able to represent more than one licence, BUT its native
REST API limits the number of provided licences to one,
and it does not have an alternate mechanism for custom licences.
So only the first licence is preserved.

Upstream has been notified about the limitation, and the potential loss
of associated licences if a metadata update occurs on entries with more
than one associated licence.