Fixed corner case caught using cache export plugin
When a dataset is exported to the input cache with a synthetic URI, the caching code was not registering the destination URI, but the intermediate one used to inject. So, the synthethic URI is never properly registered, leading to a later cache miss.

So now, cache code has a look at the returned metadata, and trusts the returned URI instead of the used for the request. This is needed for other cases like the ones where the URI encodes credentials, and the fetch plugin removes them from the URI which is embedded in the returned metadata.