Added Python lazy loading capabilities.
These capabilities are needed to delay issues with libraries like
`python-magic`, which depends on having a locatable `libmagic.so`
in the system. Loading of this library when `libmagic.so` is not
available leads to an `ImportError`, even in scenarios where no
one of its definitions is used.

This is needed to avoid having this `ImportError` when WfExS-backend
is providing the list of parameters to autodoc in Sphinx run within
the ReadTheDocs environment.