Fixed corner case in associated container images handling for CWL.
When a WRROC is generated for a CWL workflow execution, two workflows
are registered in it: the original one and the consolidated one.
The original one is used to generated the consolidated one, and the
consolidated one is used to run the workflow.

But the containers were not properly associated to the executions of this
last workflow.