Fixed several hidden issues in marshalling handling.
With the refactor of path handling along the whole codebase, some hidden issues are now arising. One of them is that marshalling and unmarshalling handling was always overwriting preoviously properly marshalled files, which should be already immutable (like stage-stage.yaml and workflow_config.yaml). Due other issues, stage-state.yaml was corrupted in case an execution fails even before the execution itself.

With execution-state.yaml, the code was "forgetting" previous executions metadata on each try.

Also, uncovered and fixed a mismatch in engine version. When a version is requested, but the selected one is different, it was partially registered in one instance variable but not in another one, so next run was doomed to unexpectedly redo the same work.