Identified race condition in FUSE mounted working directories
While trying the code within singularity, a race condition was hit on working directory creation: mount process can return even before the system has finished its mount process. This race condition does not happen outside containers.

The real solution is waiting for the mount point to be available. It seems that asking the system about a directory which is becoming a mount point in progress whether it is already one helps on that synchronization.

Added an additional error message, so if in the future this implementation does not appear to be not enough, we can realize about what it is happening, and add some polling mechanism.