Added context_inputs and context_environment to materializeWorkflow.
Some workflow engines could require to have valid inputs in order to perform a dry-run which provides introspective details about the workflow. So, both inputs and environment to be used on workflow execution are now also passed to MaterializeWorkflowAndContainers and related methods.

Also, implemented experimental inspectWorkflow method for Nextflow, to test whether it is safe to learn about the needed containers in a complex workflow.