Coherent MaterializedInput name mangling.
Some cases, like Nextflow, have structured, nested parameters. In that scenario, it is allowed that the name of the key can contain dots. As the dot is the join character being used for linearized param names, it is important to properly quote those parameters, so they are properly managed.

Also, fixed a couple of hidden bugs related to null handling and augmented parameters processing.