Implemented support for member filtering within fetched contents.
This is a new feature, which allows selecting members from fetched contents which are materialized as directories. It is a generalization of `globExplode` mechanism.

Now, support for archives has also been added, which allows extracting those members matching the defined glob patterns. This is very useful on scenarios where the dataset releases are available as .zip or .tar.gz bundles, but only some inner members are interesting.

Last, but not the least important, internal logic to place contents in the working directory has been hardened, in order to assure contents from different fetched datasets do not overwrite each other or are placed outside their boundaries.